// src/__generated__/fragmentIssueListItem_IssueFragment.graphql.ts
export type fragmentIssueListItem_IssueFragment$data = {
  readonly created_at: any;
  readonly id: any;
  readonly issue_labelsCollection: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly labels: {
          readonly color: string;
          readonly id: any;
          readonly name: string;
          readonly nodeId: string;
        } | null | undefined;
        readonly nodeId: string;
      };
    }>;
  } | null | undefined;
  readonly nodeId: string;
  readonly priority: string;
  readonly status: string;
  readonly title: string;
  readonly users: {
    readonly avatar_url: string | null | undefined;
    readonly id: any;
    readonly name: string;
    readonly nodeId: string;
  } | null | undefined;
  readonly " $fragmentType": "fragmentIssueListItem_IssueFragment";
};
export type fragmentIssueListItem_IssueFragment$key = {
  readonly " $data"?: fragmentIssueListItem_IssueFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"fragmentIssueListItem_IssueFragment">;
};